# backend/docs/openapi-module4-booking.yaml
openapi: 3.0.3
info:
  title: TURBO - Module 4 (Booking Orchestrator)
  description: |
    API contract for the Booking Orchestrator module of the TURBO peer-to-peer car rental platform.

    ## Scope
    This module handles the **complete booking lifecycle**: vehicle search, booking creation,
    owner confirmation/rejection, shift start/complete with photo documentation, and cancellation
    by either party. It connects the "supply" side (vehicles from Module 3) with the "demand"
    side (verified drivers from Module 2).

    ## Booking Flow
    1. Driver searches available vehicles via `GET /driver/bookings/vehicles/search`
    2. Driver views vehicle details via `GET /driver/bookings/vehicles/{vehicleId}`
    3. Driver creates a booking via `POST /driver/bookings`
    4. Owner confirms or rejects via `PUT /owner/bookings/{id}/confirm` or `.../reject`
    5. Driver starts shift with pickup photo via `PUT /driver/bookings/{id}/start`
    6. Driver completes shift with return photo via `PUT /driver/bookings/{id}/complete`

    ## Business Rules
    - Minimum booking duration: 4 hours
    - Maximum booking duration: 24 hours (study permit restriction)
    - Driver must be verified (`isVerified == true`) to create bookings
    - Vehicle must be APPROVED and active with `availableUntil > endTime`
    - Pickup photo required to start shift, return photo required to complete
    - Exact vehicle location revealed 2 hours before confirmed booking start time
...

    ## State Machine
...

...
servers:
...
paths:
...
  /api/driver/bookings/vehicles/search:
    get:
      tags: [DriverBooking]
      summary: Search available vehicles
      description: |
        Searches for available vehicles matching the given filters. Only vehicles with
        `status == APPROVED`, `isActive == true`, and `availableUntil > now()` are returned.
        Coordinates are masked to ~1km radius for privacy. If `startTime` and `endTime` are
        provided, vehicles with conflicting bookings (PENDING/CONFIRMED/IN_PROGRESS) are excluded.
        Geo-radius filtering uses the Haversine formula in-memory.
        Each result includes `effectiveServiceType` and `serviceTypeWarning` based on the
        driver's license class and the vehicle's service classification.
        Requires JWT authentication with DRIVER role.
      operationId: searchAvailableVehicles
      security:
        - bearerAuth: []
      parameters:
        - name: latitude
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          description: "Driver's current latitude for proximity search"
          example: 49.28
        - name: longitude
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          description: "Driver's current longitude for proximity search"
          example: -123.12
        - name: radiusKm
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 1
            maximum: 100
            default: 10.0
          description: "Search radius in kilometers (default: 10.0)"
          example: 10.0
        - name: startTime
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: "Desired shift start time. Filters out vehicles with conflicting bookings. Must be in the future."
          example: "2026-03-20T08:00:00"
        - name: endTime
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: "Desired shift end time. Must be after startTime."
          example: "2026-03-20T16:00:00"
        - name: category
          in: query
          required: false
          schema:
            type: string
            pattern: "^(SEDAN|SUV|VAN|TRUCK|COMPACT)$"
            enum: [SEDAN, SUV, VAN, TRUCK, COMPACT]
          description: "Vehicle category filter"
        - name: serviceType
          in: query
          required: false
          schema:
            type: string
            pattern: "^(TAXI_AND_DELIVERY|DELIVERY_ONLY)$"
            enum: [TAXI_AND_DELIVERY, DELIVERY_ONLY]
          description: "Service type filter (filters by vehicle classification)"
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 0
          description: "Minimum hourly rate filter"
          example: 10
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 0
          description: "Maximum hourly rate filter"
          example: 30
        - name: fuelType
          in: query
          required: false
          schema:
            type: string
            pattern: "^(GASOLINE|DIESEL|ELECTRIC|HYBRID)$"
            enum: [GASOLINE, DIESEL, ELECTRIC, HYBRID]
          description: "Fuel type filter"
      responses:
        '200':
          description: List of available vehicles matching filters (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleSearchResponse'
              example:
                - vehicleId: 5
                  make: "Toyota"
                  model: "Corolla"
                  year: 2022
                  category: "SEDAN"
                  fuelType: "GASOLINE"
                  serviceType: "TAXI_AND_DELIVERY"
                  hourlyRate: 15.50
                  description: "Well-maintained sedan, ideal for ride-sharing"
                  generalLocation: "Downtown Vancouver"
                  maskedLatitude: 49.28
                  maskedLongitude: -123.12
                  availableUntil: "2026-03-25T18:00:00"
                  ownerFullName: "Jane Smith"
                  ownerRating: 4.5
                  effectiveServiceType: "TAXI_AND_DELIVERY"
                  serviceTypeWarning: null
        '401':
          description: Not authenticated (Spring Security blocks at filter level, no JSON body)

  # ── Driver: Vehicle Detail ──────────────────────────────────────
  /api/driver/bookings/vehicles/{vehicleId}:
    get:
      tags: [DriverBooking]
      summary: View vehicle details for booking
      description: |
        Returns full vehicle details for booking consideration. Location is still masked (~1km radius).
        Includes additional fields not shown in search results: licensePlate, ownerId, vin.
        Requires JWT authentication with DRIVER role.
      operationId: getVehicleBookingDetail
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Vehicle ID
      responses:
        '200':
          description: Full vehicle details (location still masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleBookingDetailResponse'
              example:
                vehicleId: 5
                make: "Toyota"
                model: "Corolla"
                year: 2022
                category: "SEDAN"
                fuelType: "GASOLINE"
                serviceType: "TAXI_AND_DELIVERY"
                hourlyRate: 15.50
                description: "Well-maintained sedan, ideal for ride-sharing"
                generalLocation: "Downtown Vancouver"
                maskedLatitude: 49.28
                maskedLongitude: -123.12
                availableUntil: "2026-03-25T18:00:00"
                ownerFullName: "Jane Smith"
                ownerRating: 4.5
                effectiveServiceType: "TAXI_AND_DELIVERY"
                serviceTypeWarning: null
                licensePlate: "ABC 1234"
                ownerId: 3
                vin: "1HGBH41JXMN109186"
        '401':
          description: Not authenticated
        '404':
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VEH-004
                message: "Vehicle not found"
                status: 404
                timestamp: "2026-03-17T10:00:00.000"

  # ── Driver: Create Booking ──────────────────────────────────────
  /api/driver/bookings:
    post:
      tags: [DriverBooking]
      summary: Create a new booking
      description: |
        Creates a new booking for the specified vehicle and time range. The driver must be verified.
        The system validates: vehicle availability, time constraints (min 4h, max 24h, future start),
        no overlapping bookings for the vehicle or the driver, and that the driver is not the vehicle owner.
        Price is calculated as `totalHours * vehicle.hourlyRate` and stored at creation time.
        Booking starts in PENDING status awaiting owner confirmation.
        An email notification is sent to the vehicle owner.
        Requires JWT authentication with DRIVER role.
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              vehicleId: 5
              startTime: "2026-03-20T08:00:00"
              endTime: "2026-03-20T16:00:00"
      responses:
        '200':
          description: Booking created successfully with status PENDING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                bookingId: 1
                driverId: 2
                driverFullName: "John Doe"
                vehicleId: 5
                vehicleSummary: "2022 Toyota Corolla"
                status: "PENDING"
                startTime: "2026-03-20T08:00:00"
                endTime: "2026-03-20T16:00:00"
                totalHours: 8
                totalPrice: 124.00
                createdAt: "2026-03-17T10:00:00.000"
        '400':
          description: Validation or business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                vehicleNotApproved:
                  summary: Vehicle not approved (BOOK-002)
                  value:
                    code: BOOK-002
                    message: "Vehicle is not approved for rentals"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                vehicleNotActive:
                  summary: Vehicle not active (BOOK-003)
                  value:
                    code: BOOK-003
                    message: "Vehicle is not currently available for rent"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                exceedsAvailability:
                  summary: End time exceeds vehicle availability (BOOK-004)
                  value:
                    code: BOOK-004
                    message: "Booking end time exceeds vehicle availability window"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                startNotFuture:
                  summary: Start time not in future (BOOK-005)
                  value:
                    code: BOOK-005
                    message: "Booking start time must be in the future"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                endBeforeStart:
                  summary: End time before start time (BOOK-006)
                  value:
                    code: BOOK-006
                    message: "Booking end time must be after start time"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                tooShort:
                  summary: Duration less than 4 hours (BOOK-007)
                  value:
                    code: BOOK-007
                    message: "Minimum booking duration is 4 hours"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                tooLong:
                  summary: Duration exceeds 24 hours (BOOK-025)
                  value:
                    code: BOOK-025
                    message: "Maximum booking duration is 24 hours"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                ownVehicle:
                  summary: Cannot book own vehicle (BOOK-009)
                  value:
                    code: BOOK-009
                    message: "You cannot book your own vehicle"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                validationError:
                  summary: Field validation errors (VALIDATION-001)
                  value:
                    code: VALIDATION-001
                    error: "Validation Failed"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                    details:
                      vehicleId: "Vehicle ID is required"
                      startTime: "Start time is required"
        '403':
          description: Driver not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-001
                message: "Driver must be verified before creating bookings"
                status: 403
                timestamp: "2026-03-17T10:00:00.000"
        '401':
          description: Not authenticated
        '404':
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VEH-004
                message: "Vehicle not found"
                status: 404
                timestamp: "2026-03-17T10:00:00.000"
        '409':
          description: Time slot conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                vehicleConflict:
                  summary: Vehicle already booked (BOOK-008)
                  value:
                    code: BOOK-008
                    message: "Vehicle is already booked for the requested time slot"
                    status: 409
                    timestamp: "2026-03-17T10:00:00.000"
                driverConflict:
                  summary: Driver has overlapping booking (BOOK-010)
                  value:
                    code: BOOK-010
                    message: "You already have an active booking during this time slot"
                    status: 409
                    timestamp: "2026-03-17T10:00:00.000"

    # ── Driver: List My Bookings ────────────────────────────────────

    get:
      tags: [DriverBooking]
      summary: List my bookings (as driver)
      description: |
        Returns all bookings created by the authenticated driver. Optionally filter by booking status.
        Requires JWT authentication with DRIVER role.
      operationId: listDriverBookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            pattern: "^(PENDING|CONFIRMED|IN_PROGRESS|COMPLETED|CANCELLED|REJECTED)$"
            enum: [PENDING, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED, REJECTED]
          description: "Filter bookings by status (optional)"
      responses:
        '200':
          description: List of driver's bookings (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingResponse'
              example:
                - bookingId: 1
                  driverId: 2
                  driverFullName: "John Doe"
                  vehicleId: 5
                  vehicleSummary: "2022 Toyota Corolla"
                  status: "CONFIRMED"
                  startTime: "2026-03-20T08:00:00"
                  endTime: "2026-03-20T16:00:00"
                  totalHours: 8
                  totalPrice: 124.00
                  createdAt: "2026-03-17T10:00:00.000"
                - bookingId: 2
                  driverId: 2
                  driverFullName: "John Doe"
                  vehicleId: 8
                  vehicleSummary: "2021 Honda Civic"
                  status: "PENDING"
                  startTime: "2026-03-22T09:00:00"
                  endTime: "2026-03-22T17:00:00"
                  totalHours: 8
                  totalPrice: 96.00
                  createdAt: "2026-03-17T11:00:00.000"
        '401':
          description: Not authenticated

  # ── Driver: Get Booking Detail ──────────────────────────────────
  /api/driver/bookings/{bookingId}:
    get:
      tags: [DriverBooking]
      summary: Get booking details (driver view)
      description: |
        Returns full booking details for the authenticated driver. Includes vehicle info, owner info,
        photos, cancellation details, and timestamps. Location is masked unless within 2 hours of
        a CONFIRMED/IN_PROGRESS booking start time.
        Requires JWT authentication with DRIVER role.
      operationId: getDriverBookingDetail
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Booking ID
      responses:
        '200':
          description: Full booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
              example:
                bookingId: 1
                driverId: 2
                driverFullName: "John Doe"
                vehicleId: 5
                vehicleSummary: "2022 Toyota Corolla"
                status: "CONFIRMED"
                startTime: "2026-03-20T08:00:00"
                endTime: "2026-03-20T16:00:00"
                totalHours: 8
                totalPrice: 124.00
                createdAt: "2026-03-17T10:00:00.000"
                ownerId: 3
                ownerFullName: "Jane Smith"
                vehicleMake: "Toyota"
                vehicleModel: "Corolla"
                vehicleYear: 2022
                vehicleLicensePlate: "ABC 1234"
                vehicleCategory: "SEDAN"
                vehicleServiceType: "TAXI_AND_DELIVERY"
                vehicleHourlyRate: 15.50
                pickupLocation: "Downtown Vancouver"
                pickupPhotoUrls: []
                returnPhotoUrls: []
                cancellationReason: null
                cancelledBy: null
                confirmedAt: "2026-03-17T12:00:00.000"
                startedAt: null
                completedAt: null
                cancelledAt: null
        '401':
          description: Not authenticated
        '403':
          description: Booking does not belong to the authenticated driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-012
                message: "You do not have permission to access this booking"
                status: 403
                timestamp: "2026-03-17T10:00:00.000"
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-011
                message: "Booking not found"
                status: 404
                timestamp: "2026-03-17T10:00:00.000"

  # ── Driver: Cancel Booking ──────────────────────────────────────
  /api/driver/bookings/{bookingId}/cancel:
    put:
      tags: [DriverBooking]
      summary: Cancel a pending or confirmed booking (driver)
      description: |
        Cancels a booking that is in PENDING or CONFIRMED status. A cancellation reason is required.
        Sets `cancelledBy = "DRIVER"`. An email notification is sent to the vehicle owner.
        Requires JWT authentication with DRIVER role.
      operationId: cancelBookingByDriver
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Booking ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
            example:
              reason: "Schedule conflict, cannot make it"
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                bookingId: 1
                driverId: 2
                driverFullName: "John Doe"
                vehicleId: 5
                vehicleSummary: "2022 Toyota Corolla"
                status: "CANCELLED"
                startTime: "2026-03-20T08:00:00"
                endTime: "2026-03-20T16:00:00"
                totalHours: 8
                totalPrice: 124.00
                createdAt: "2026-03-17T10:00:00.000"
        '400':
          description: Invalid status or missing reason
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidStatus:
                  summary: Booking not in cancellable status (BOOK-015)
                  value:
                    code: BOOK-015
                    message: "Only pending or confirmed bookings can be cancelled"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
                missingReason:
                  summary: Missing cancellation reason (BOOK-022)
                  value:
                    code: BOOK-022
                    message: "Cancellation reason is required"
                    status: 400
                    timestamp: "2026-03-17T10:00:00.000"
        '401':
          description: Not authenticated
        '403':
          description: Booking does not belong to the authenticated driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-012
                message: "You do not have permission to access this booking"
                status: 403
                timestamp: "2026-03-17T10:00:00.000"
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-011
                message: "Booking not found"
                status: 404
                timestamp: "2026-03-17T10:00:00.000"

  # ── Driver: Start Shift ─────────────────────────────────────────
...
  /api/driver/bookings/{bookingId}/vehicle-location:
    get:
      tags: [DriverBooking]
      summary: Get vehicle location (masked or exact)
      description: |
        Returns the vehicle pickup location for the booking. The location is masked (~1km radius)
        by default and becomes exact when the booking is CONFIRMED/IN_PROGRESS and within 2 hours
        of the start time. For COMPLETED bookings, exact location is returned for reference.
        Only valid for bookings in CONFIRMED, IN_PROGRESS, or COMPLETED status.
        Requires JWT authentication with DRIVER role.
      operationId: getVehicleLocation
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Booking ID
      responses:
        '200':
          description: Vehicle location (masked or exact depending on timing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleLocationResponse'
              examples:
                masked:
                  summary: Masked location (more than 2h before start)
                  value:
                    vehicleId: 5
                    isExactLocation: false
                    generalLocation: "Downtown Vancouver"
                    latitude: 49.28
                    longitude: -123.12
                    message: "Exact location will be available 2 hours before your booking"
                exact:
                  summary: Exact location (within 2h of start)
                  value:
                    vehicleId: 5
                    isExactLocation: true
                    generalLocation: "Downtown Vancouver"
                    latitude: 49.2827
                    longitude: -123.1207
                    message: "Exact pickup location is now available"
        '400':
          description: Booking not in valid status for location reveal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-016
                message: "Only confirmed bookings can be started"
                status: 400
                timestamp: "2026-03-17T10:00:00.000"
        '401':
          description: Not authenticated
        '403':
          description: Booking does not belong to the authenticated driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-012
                message: "You do not have permission to access this booking"
                status: 403
                timestamp: "2026-03-17T10:00:00.000"
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BOOK-011
                message: "Booking not found"
                status: 404
                timestamp: "2026-03-17T10:00:00.000"

  # ── Owner: List Bookings ────────────────────────────────────────
...
components:
  schemas:
...
    VehicleBookingDetailResponse:
      type: object
      description: "Full vehicle details for booking consideration. Extends VehicleSearchResponse with licensePlate, ownerId, and vin."
      allOf:
        - $ref: '#/components/schemas/VehicleSearchResponse'
        - type: object
          properties:
            licensePlate:
              type: string
              description: "Vehicle license plate"
              example: "ABC 1234"
            ownerId:
              type: integer
              format: int64
              description: "Vehicle owner user ID"
              example: 3
            vin:
              type: string
              description: "Vehicle Identification Number"
              example: "1HGBH41JXMN109186"
...
    ErrorResponse:
      type: object
      description: "Standard business error response. Used for all BOOK-XXX, VEH-XXX, DOC-XXX, and AUTH-XXX errors."
      properties:
        code:
          type: string
          description: "Error code (e.g. BOOK-001, VEH-004)"
          example: "BOOK-011"
        message:
          type: string
          description: "Human-readable error message"
          example: "Booking not found"
        status:
          type: integer
          description: "HTTP status code"
          example: 404
        timestamp:
          type: string
          format: date-time
          description: "Error timestamp (ISO-8601)"
          example: "2026-03-17T10:00:00.000"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via POST /api/auth/login or POST /api/auth/verify-otp (Module 1).
        Token payload includes: userId, email, role (DRIVER, CAR_OWNER, or ADMIN).
        Must be sent in the Authorization header: `Bearer <token>`.